requirements.txt
alembic==1.14.0
amqp==5.3.1
anyio==4.6.2.post1
astroid==2.13.5
async-timeout==5.0.1
asyncio==3.4.3
attrs==24.2.0
bcrypt==4.0.1
billiard==3.6.4.0
cachetools==5.5.0
celery==5.2.7
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dill==0.3.9
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.88.0
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
h11==0.14.0
httpcore==0.16.3
httplib2==0.22.0
httptools==0.6.4
httpx==0.23.1
idna==3.10
iniconfig==2.0.0
isort==5.13.2
kombu==5.4.2
lazy-object-proxy==1.10.0
loguru==0.6.0
Mako==1.3.6
MarkupSafe==3.0.2
mccabe==0.7.0
oauthlib==3.2.2
packaging==24.2
passlib==1.7.4
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.29.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==1.10.19
pylint==2.15.9
pymongo==4.3.3
PyMySQL==1.1.1
pyparsing==3.2.0
pytest==7.2.0
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.2
PyYAML==6.0.2
redis==4.4.0
requests==2.32.3
requests-oauthlib==2.0.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
spotipy==2.24.0
SQLAlchemy==1.4.45
sse-starlette==1.2.1
starlette==0.22.0
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.20.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.0
wcwidth==0.2.13
websockets==14.1
wrapt==1.17.0
